Update our Unicode and legacy codepage support to Unicode 16.0.0
We missed a couple of releases so this adds 11,131 new characters
compared to our previous version 13.0.0.  These are mostly obscure
alphabets and emojis (I'm just guessing) that are not super likely
to be included in any password.

There were also a few minor changes to definitions of categories such
as punctuation and specials - we always used the Unicode Database's
definition of them and apparently they changed their mind about a few.

Closes #5537