OpenCL: Double the performance of shared AES-CBC-decrypt
Ensure AES_cbc_decrypt() utilizes the parallel capability of the bitsliced
AES code when more than one block is decrypted at once.  For encryption this
is not possible as the IV for block n is unknown until block n-1 has been
encrypted.

The potential boost is great (over 2x seen with AES-ECB) but none of our
current formats use AES-CBC enough for this to make a significant
difference, at least not with their (often small) test vectors.

See #5594.